Remove TH (#16)
* Remove TH - no longer needed

* Consolidate FixedVector into a single module
