+++ riscv-hs.cabal
@@ -24,9 +24,6 @@ library
         FixedVector,
         Register
     other-modules:
-        FixedVector.TH,
-        FixedVector.Type,
-        FixedVector.Constructors
     -- other-extensions:
     build-depends:    base, template-haskell
     hs-source-dirs:   src
